Update dock and names.
